Refactor buildIndex functions into separate modules
- Extract extractCaseSummary function into its own module
- Extract extractDocuments function into its own module
- Create dedicated test files for each extracted function
- Update buildIndex.ts to import from new modules
- Remove combined buildIndex.test.ts file
- Fix npm test script to run without watch mode

This refactoring prevents the buildIndex main function from executing during tests
and improves code organization by separating concerns.
